Change namespace to `no.novari.flyt.discovery.service.*`
