--- src/main/java/no/novari/discovery/service/mapping/InstanceValueMetadataMappingService.java
@@ -1,28 +1,0 @@
-package no.novari.discovery.service.mapping;
-
-import no.novari.discovery.service.model.dtos.InstanceValueMetadataDto;
-import no.novari.discovery.service.model.entities.InstanceValueMetadata;
-import org.springframework.stereotype.Service;
-
-@Service
-public class InstanceValueMetadataMappingService {
-
-    public InstanceValueMetadata toEntity(InstanceValueMetadataDto instanceValueMetadataDto) {
-        return InstanceValueMetadata
-                .builder()
-                .displayName(instanceValueMetadataDto.getDisplayName())
-                .type(instanceValueMetadataDto.getType())
-                .key(instanceValueMetadataDto.getKey())
-                .build();
-    }
-
-    public InstanceValueMetadataDto toDto(InstanceValueMetadata instanceValueMetadata) {
-        return InstanceValueMetadataDto
-                .builder()
-                .displayName(instanceValueMetadata.getDisplayName())
-                .type(instanceValueMetadata.getType())
-                .key(instanceValueMetadata.getKey())
-                .build();
-    }
-
-}
