--- src/main/java/no/novari/discovery/service/mapping/IntegrationMetadataMappingService.java
@@ -1,45 +1,0 @@
-package no.novari.discovery.service.mapping;
-
-import no.novari.discovery.service.model.dtos.IntegrationMetadataDto;
-import no.novari.discovery.service.model.entities.IntegrationMetadata;
-import org.springframework.stereotype.Service;
-
-@Service
-public class IntegrationMetadataMappingService {
-
-    private final InstanceMetadataContentMappingService instanceMetadataContentMappingService;
-
-    public IntegrationMetadataMappingService(InstanceMetadataContentMappingService instanceMetadataContentMappingService) {
-        this.instanceMetadataContentMappingService = instanceMetadataContentMappingService;
-    }
-
-    public IntegrationMetadata toEntity(IntegrationMetadataDto integrationMetadataDto) {
-        return IntegrationMetadata
-                .builder()
-                .sourceApplicationId(integrationMetadataDto.getSourceApplicationId())
-                .sourceApplicationIntegrationId(integrationMetadataDto.getSourceApplicationIntegrationId())
-                .sourceApplicationIntegrationUri(integrationMetadataDto.getSourceApplicationIntegrationUri())
-                .integrationDisplayName(integrationMetadataDto.getIntegrationDisplayName())
-                .version(integrationMetadataDto.getVersion())
-                .instanceMetadata(
-                        instanceMetadataContentMappingService.toEntity(integrationMetadataDto.getInstanceMetadata())
-                )
-                .build();
-    }
-
-    public IntegrationMetadataDto toDto(IntegrationMetadata integrationMetadata) {
-        return IntegrationMetadataDto
-                .builder()
-                .id(integrationMetadata.getId())
-                .sourceApplicationId(integrationMetadata.getSourceApplicationId())
-                .sourceApplicationIntegrationId(integrationMetadata.getSourceApplicationIntegrationId())
-                .sourceApplicationIntegrationUri(integrationMetadata.getSourceApplicationIntegrationUri())
-                .integrationDisplayName(integrationMetadata.getIntegrationDisplayName())
-                .version(integrationMetadata.getVersion())
-                .instanceMetadata(
-                        instanceMetadataContentMappingService.toDto(integrationMetadata.getInstanceMetadata())
-                )
-                .build();
-    }
-
-}
