--- src/main/java/no/novari/discovery/service/model/entities/InstanceMetadataContent.java
@@ -1,61 +1,0 @@
-package no.novari.discovery.service.model.entities;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.extern.jackson.Jacksonized;
-
-import java.util.Collection;
-import java.util.Objects;
-
-@Getter
-@Setter
-@Jacksonized
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Entity
-public class InstanceMetadataContent {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @JsonIgnore
-    @Setter(AccessLevel.NONE)
-    private long id;
-
-    @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
-    private Collection<@Valid @NotNull InstanceValueMetadata> instanceValueMetadata;
-
-    @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
-    private Collection<@Valid @NotNull InstanceObjectCollectionMetadata> instanceObjectCollectionMetadata;
-
-    @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
-    private Collection<@Valid @NotNull InstanceMetadataCategory> categories;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        InstanceMetadataContent that = (InstanceMetadataContent) o;
-        return id == that.id;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-
-
-}
