--- src/main/java/no/novari/discovery/service/model/entities/InstanceObjectCollectionMetadata.java
@@ -1,58 +1,0 @@
-package no.novari.discovery.service.model.entities;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToOne;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.extern.jackson.Jacksonized;
-
-import java.util.Objects;
-
-@Getter
-@Setter
-@Jacksonized
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Entity
-public class InstanceObjectCollectionMetadata {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @JsonIgnore
-    @Setter(AccessLevel.NONE)
-    private long id;
-
-    private String displayName;
-
-    @OneToOne(cascade = CascadeType.ALL, orphanRemoval = true)
-    private InstanceMetadataContent objectMetadata;
-
-    @Column(name = "\"key\"")
-    private String key;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        InstanceObjectCollectionMetadata that = (InstanceObjectCollectionMetadata) o;
-        return id == that.id;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-
-
-}
