+++ src/main/java/no/novari/flyt/discovery/service/model/entities/InstanceValueMetadata.java
@@ -1,0 +1,63 @@
+package no.novari.flyt.discovery.service.model.entities;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.extern.jackson.Jacksonized;
+
+import java.util.Objects;
+
+@Getter
+@Setter
+@Jacksonized
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Entity
+public class InstanceValueMetadata {
+
+    public enum Type {
+        STRING,
+        BOOLEAN,
+        FILE
+    }
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @JsonIgnore
+    @Setter(AccessLevel.NONE)
+    private long id;
+
+    private String displayName;
+
+    @Enumerated(EnumType.STRING)
+    private Type type;
+
+    @Column(name = "\"key\"")
+    private String key;
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        InstanceValueMetadata that = (InstanceValueMetadata) o;
+        return id == that.id;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
+    }
+
+}
