+++ src/test/java/no/novari/flyt/discovery/service/mapping/IntegrationMetadataMappingIntegrationTest.java
@@ -1,0 +1,334 @@
+package no.novari.flyt.discovery.service.mapping;
+
+import no.novari.flyt.discovery.service.model.dtos.InstanceMetadataCategoryDto;
+import no.novari.flyt.discovery.service.model.dtos.InstanceMetadataContentDto;
+import no.novari.flyt.discovery.service.model.dtos.InstanceObjectCollectionMetadataDto;
+import no.novari.flyt.discovery.service.model.dtos.InstanceValueMetadataDto;
+import no.novari.flyt.discovery.service.model.dtos.IntegrationMetadataDto;
+import no.novari.flyt.discovery.service.model.entities.InstanceMetadataCategory;
+import no.novari.flyt.discovery.service.model.entities.InstanceMetadataContent;
+import no.novari.flyt.discovery.service.model.entities.InstanceObjectCollectionMetadata;
+import no.novari.flyt.discovery.service.model.entities.InstanceValueMetadata;
+import no.novari.flyt.discovery.service.model.entities.IntegrationMetadata;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import java.util.List;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+@SpringBootTest(classes = {
+        IntegrationMetadataMappingService.class,
+        InstanceMetadataContentMappingService.class,
+        InstanceValueMetadataMappingService.class,
+        InstanceObjectCollectionMetadataMappingService.class,
+        InstanceMetadataCategoryMappingService.class
+})
+public class IntegrationMetadataMappingIntegrationTest {
+
+    @Autowired
+    private IntegrationMetadataMappingService integrationMetadataMappingService;
+
+    private IntegrationMetadataDto integrationMetadataDto;
+    private IntegrationMetadata expectedIntegrationMetadata;
+
+    @BeforeEach
+    public void setup() {
+        integrationMetadataDto = IntegrationMetadataDto
+                .builder()
+                .sourceApplicationId(1L)
+                .sourceApplicationIntegrationId("Test-1")
+                .version(1L)
+                .integrationDisplayName("displayName")
+                .instanceMetadata(
+                        InstanceMetadataContentDto
+                                .builder()
+                                .instanceValueMetadata(List.of(
+                                        InstanceValueMetadataDto
+                                                .builder()
+                                                .displayName("value1DisplayName")
+                                                .type(InstanceValueMetadata.Type.STRING)
+                                                .key("value1Key")
+                                                .build(),
+                                        InstanceValueMetadataDto
+                                                .builder()
+                                                .displayName("value2DisplayName")
+                                                .type(InstanceValueMetadata.Type.STRING)
+                                                .key("value2Key")
+                                                .build()
+                                ))
+                                .instanceObjectCollectionMetadata(List.of(
+                                        InstanceObjectCollectionMetadataDto
+                                                .builder()
+                                                .displayName("objectCollection1DisplayName")
+                                                .objectMetadata(
+                                                        InstanceMetadataContentDto
+                                                                .builder()
+                                                                .instanceValueMetadata(List.of(
+                                                                        InstanceValueMetadataDto
+                                                                                .builder()
+                                                                                .displayName("objectCollection1value1DisplayName")
+                                                                                .type(InstanceValueMetadata.Type.BOOLEAN)
+                                                                                .key("objectCollection1value1Key")
+                                                                                .build()
+                                                                ))
+                                                                .instanceObjectCollectionMetadata(List.of())
+                                                                .categories(List.of())
+                                                                .build()
+                                                )
+                                                .key("objectCollection1key")
+                                                .build(),
+                                        InstanceObjectCollectionMetadataDto
+                                                .builder()
+                                                .displayName("objectCollection2DisplayName")
+                                                .objectMetadata(
+                                                        InstanceMetadataContentDto
+                                                                .builder()
+                                                                .instanceValueMetadata(List.of(
+                                                                        InstanceValueMetadataDto
+                                                                                .builder()
+                                                                                .displayName("objectCollection2value1DisplayName")
+                                                                                .type(InstanceValueMetadata.Type.BOOLEAN)
+                                                                                .key("objectCollection2value1Key")
+                                                                                .build(),
+                                                                        InstanceValueMetadataDto
+                                                                                .builder()
+                                                                                .displayName("objectCollection2value2DisplayName")
+                                                                                .type(InstanceValueMetadata.Type.FILE)
+                                                                                .key("objectCollection2value2Key")
+                                                                                .build()
+                                                                ))
+                                                                .instanceObjectCollectionMetadata(List.of(
+                                                                        InstanceObjectCollectionMetadataDto
+                                                                                .builder()
+                                                                                .displayName("objectCollection2objectCollection1DisplayName")
+                                                                                .objectMetadata(
+                                                                                        InstanceMetadataContentDto
+                                                                                                .builder()
+                                                                                                .instanceValueMetadata(List.of(
+                                                                                                        InstanceValueMetadataDto
+                                                                                                                .builder()
+                                                                                                                .displayName("objectCollection2objectCollection1value1DisplayName")
+                                                                                                                .type(InstanceValueMetadata.Type.STRING)
+                                                                                                                .key("objectCollection2objectCollection1value1Key")
+                                                                                                                .build()
+                                                                                                ))
+                                                                                                .build()
+                                                                                )
+                                                                                .key("objectCollection2objectCollection1key")
+                                                                                .build()
+                                                                ))
+                                                                .build()
+                                                )
+                                                .key("objectCollection2key")
+                                                .build()
+                                ))
+                                .categories(List.of(
+                                        InstanceMetadataCategoryDto
+                                                .builder()
+                                                .displayName("category1Key")
+                                                .content(
+                                                        InstanceMetadataContentDto
+                                                                .builder()
+                                                                .categories(List.of(
+                                                                        InstanceMetadataCategoryDto
+                                                                                .builder()
+                                                                                .displayName("category1category1Key")
+                                                                                .content(
+                                                                                        InstanceMetadataContentDto
+                                                                                                .builder()
+                                                                                                .instanceValueMetadata(List.of(
+                                                                                                        InstanceValueMetadataDto
+                                                                                                                .builder()
+                                                                                                                .displayName("category1category1value1DisplayName")
+                                                                                                                .type(InstanceValueMetadata.Type.STRING)
+                                                                                                                .key("category1category2value1Key")
+                                                                                                                .build()
+                                                                                                ))
+                                                                                                .build()
+                                                                                )
+                                                                                .build()
+                                                                ))
+                                                                .build()
+                                                )
+                                                .build(),
+                                        InstanceMetadataCategoryDto
+                                                .builder()
+                                                .displayName("category2Key")
+                                                .content(
+                                                        InstanceMetadataContentDto
+                                                                .builder()
+                                                                .instanceValueMetadata(List.of(
+                                                                        InstanceValueMetadataDto
+                                                                                .builder()
+                                                                                .displayName("category2value1DisplayName")
+                                                                                .type(InstanceValueMetadata.Type.STRING)
+                                                                                .key("category2value1Key")
+                                                                                .build()
+                                                                ))
+                                                                .build()
+                                                )
+                                                .build()
+                                ))
+                                .build()
+                )
+                .build();
+
+        expectedIntegrationMetadata = IntegrationMetadata
+                .builder()
+                .sourceApplicationId(1L)
+                .sourceApplicationIntegrationId("Test-1")
+                .version(1L)
+                .integrationDisplayName("displayName")
+                .instanceMetadata(
+                        InstanceMetadataContent
+                                .builder()
+                                .instanceValueMetadata(List.of(
+                                        InstanceValueMetadata
+                                                .builder()
+                                                .displayName("value1DisplayName")
+                                                .type(InstanceValueMetadata.Type.STRING)
+                                                .key("value1Key")
+                                                .build(),
+                                        InstanceValueMetadata
+                                                .builder()
+                                                .displayName("value2DisplayName")
+                                                .type(InstanceValueMetadata.Type.STRING)
+                                                .key("value2Key")
+                                                .build()
+                                ))
+                                .instanceObjectCollectionMetadata(List.of(
+                                        InstanceObjectCollectionMetadata
+                                                .builder()
+                                                .displayName("objectCollection1DisplayName")
+                                                .objectMetadata(
+                                                        InstanceMetadataContent
+                                                                .builder()
+                                                                .instanceValueMetadata(List.of(
+                                                                        InstanceValueMetadata
+                                                                                .builder()
+                                                                                .displayName("objectCollection1value1DisplayName")
+                                                                                .type(InstanceValueMetadata.Type.BOOLEAN)
+                                                                                .key("objectCollection1value1Key")
+                                                                                .build()
+                                                                ))
+                                                                .instanceObjectCollectionMetadata(List.of())
+                                                                .categories(List.of())
+                                                                .build()
+                                                )
+                                                .key("objectCollection1key")
+                                                .build(),
+                                        InstanceObjectCollectionMetadata
+                                                .builder()
+                                                .displayName("objectCollection2DisplayName")
+                                                .objectMetadata(
+                                                        InstanceMetadataContent
+                                                                .builder()
+                                                                .instanceValueMetadata(List.of(
+                                                                        InstanceValueMetadata
+                                                                                .builder()
+                                                                                .displayName("objectCollection2value1DisplayName")
+                                                                                .type(InstanceValueMetadata.Type.BOOLEAN)
+                                                                                .key("objectCollection2value1Key")
+                                                                                .build(),
+                                                                        InstanceValueMetadata
+                                                                                .builder()
+                                                                                .displayName("objectCollection2value2DisplayName")
+                                                                                .type(InstanceValueMetadata.Type.FILE)
+                                                                                .key("objectCollection2value2Key")
+                                                                                .build()
+                                                                ))
+                                                                .instanceObjectCollectionMetadata(List.of(
+                                                                        InstanceObjectCollectionMetadata
+                                                                                .builder()
+                                                                                .displayName("objectCollection2objectCollection1DisplayName")
+                                                                                .objectMetadata(
+                                                                                        InstanceMetadataContent
+                                                                                                .builder()
+                                                                                                .instanceValueMetadata(List.of(
+                                                                                                        InstanceValueMetadata
+                                                                                                                .builder()
+                                                                                                                .displayName("objectCollection2objectCollection1value1DisplayName")
+                                                                                                                .type(InstanceValueMetadata.Type.STRING)
+                                                                                                                .key("objectCollection2objectCollection1value1Key")
+                                                                                                                .build()
+                                                                                                ))
+                                                                                                .instanceObjectCollectionMetadata(List.of())
+                                                                                                .categories(List.of())
+                                                                                                .build()
+                                                                                )
+                                                                                .key("objectCollection2objectCollection1key")
+                                                                                .build()
+                                                                ))
+                                                                .categories(List.of())
+                                                                .build()
+                                                )
+                                                .key("objectCollection2key")
+                                                .build()
+                                ))
+                                .categories(List.of(
+                                        InstanceMetadataCategory
+                                                .builder()
+                                                .displayName("category1Key")
+                                                .content(
+                                                        InstanceMetadataContent
+                                                                .builder()
+                                                                .instanceValueMetadata(List.of())
+                                                                .instanceObjectCollectionMetadata(List.of())
+                                                                .categories(List.of(
+                                                                        InstanceMetadataCategory
+                                                                                .builder()
+                                                                                .displayName("category1category1Key")
+                                                                                .content(
+                                                                                        InstanceMetadataContent
+                                                                                                .builder()
+                                                                                                .instanceValueMetadata(List.of(
+                                                                                                        InstanceValueMetadata
+                                                                                                                .builder()
+                                                                                                                .displayName("category1category1value1DisplayName")
+                                                                                                                .type(InstanceValueMetadata.Type.STRING)
+                                                                                                                .key("category1category2value1Key")
+                                                                                                                .build()
+                                                                                                ))
+                                                                                                .instanceObjectCollectionMetadata(List.of())
+                                                                                                .categories(List.of())
+                                                                                                .build()
+                                                                                )
+                                                                                .build()
+                                                                ))
+                                                                .build()
+                                                )
+                                                .build(),
+                                        InstanceMetadataCategory
+                                                .builder()
+                                                .displayName("category2Key")
+                                                .content(
+                                                        InstanceMetadataContent
+                                                                .builder()
+                                                                .instanceValueMetadata(List.of(
+                                                                        InstanceValueMetadata
+                                                                                .builder()
+                                                                                .displayName("category2value1DisplayName")
+                                                                                .type(InstanceValueMetadata.Type.STRING)
+                                                                                .key("category2value1Key")
+                                                                                .build()
+                                                                ))
+                                                                .instanceObjectCollectionMetadata(List.of())
+                                                                .categories(List.of())
+                                                                .build()
+                                                )
+                                                .build()
+                                ))
+                                .build()
+                )
+                .build();
+    }
+
+    @Test
+    public void shouldMapToIntegrationMetadata() {
+        IntegrationMetadata integrationMetadata = integrationMetadataMappingService.toEntity(integrationMetadataDto);
+        assertEquals(expectedIntegrationMetadata, integrationMetadata);
+    }
+}
